% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_cobalt_installation.R
\name{check_cobalt_installation}
\alias{check_cobalt_installation}
\title{Check that COBALT is installed}
\usage{
check_cobalt_installation(
  cobalt_folder = file.path(rappdirs::user_data_dir(), "cobalt")
)
}
\description{
Check that COBALT is installed, give a helpful error message if not
}
