% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_cobalt_options.R
\name{check_cobalt_options}
\alias{check_cobalt_options}
\title{Check if the argument is a valid \code{cobalt_options}}
\usage{
check_cobalt_options(cobalt_options)
}
\arguments{
\item{cobalt_options}{a set of COBALT options,
as can be created by \link{create_cobalt_options}}
}
\description{
Will \link{stop} with a helpful error message if not
}
