% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_cobalt_options.R
\name{check_cobalt_options_booleans}
\alias{check_cobalt_options_booleans}
\title{Internal function}
\usage{
check_cobalt_options_booleans(cobalt_options)
}
\arguments{
\item{cobalt_options}{a set of COBALT options,
as can be created by \link{create_cobalt_options}}
}
\description{
Check if the booleans in the \code{cobalt_options}
are of the right datatype
}
