% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cobaltr.R
\docType{package}
\name{cobaltr}
\alias{cobaltr}
\title{\code{cobaltr}: Interface to COBALT.}
\description{
\code{cobaltr} allows to call COBALT, a popular
multiple alignment tool, using
an R interface. 'cobaltr' closely follows the interface
of COBALT, including its default settings.
}
