% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cobaltr_report.R
\name{cobaltr_report}
\alias{cobaltr_report}
\title{Create a \link{cobaltr} report, to be used when reporting bugs}
\usage{
cobaltr_report()
}
\description{
Create a \link{cobaltr} report, to be used when reporting bugs
}
