% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_cobalt_args.R
\name{create_cobalt_args}
\alias{create_cobalt_args}
\title{Create the command arguments to call COBALT with}
\usage{
create_cobalt_args(fasta_filename, cobalt_options)
}
\arguments{
\item{fasta_filename}{name of a FASTA file}

\item{cobalt_options}{a set of COBALT options,
as can be created by \link{create_cobalt_options}}
}
\value{
the command and its arguments
}
\description{
Create the command arguments to call COBALT with
}
