% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_cobalt_options.R
\name{create_cobalt_options}
\alias{create_cobalt_options}
\title{Create the COBALT options}
\usage{
create_cobalt_options(
  search_pattern_filename = file.path(get_default_cobalt_folder(), "patterns"),
  domain_database_filename = file.path(get_default_cobalt_folder(), "cdd"),
  conserved_block_filename = file.path(get_default_cobalt_folder(), "cdd.blocks"),
  residue_frequencies_filename = file.path(get_default_cobalt_folder(), "cdd.freq"),
  destination_filename = "",
  pairwise_alignment_constraints_filename = "",
  conserved_domain_threshold = 0.01,
  filler_threshold = 0.01,
  terminal_open_gap_penalty = 5,
  terminal_extend_gap_penalty = 1,
  open_gap_in_middle_penalty = 11,
  middle_extend_gap_penalty = 1,
  conservation_score_threshold = 0.67,
  dfb_added_weight = 0.5,
  ffb_added_weight = 1,
  no_rps = FALSE,
  search_around_conserved_columns = TRUE,
  score_matrix_name = "BLOSUM62",
  pseudocount_constant = 2,
  use_fastme = FALSE,
  do_dry_run = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{search_pattern_filename}{filename containing search patterns,
the COBALT \code{-p} option}

\item{domain_database_filename}{domain database filename,
the COBALT \code{-db} option}

\item{conserved_block_filename}{filename containing conserved blocks,
the COBALT \code{-b} option}

\item{residue_frequencies_filename}{filename containing residue frequencies,
the COBALT \code{-f} option}

\item{destination_filename}{filename for destination text seqalign
the COBALT \code{-seqalign} option}

\item{pairwise_alignment_constraints_filename}{name of the file
containing pairwise alignment constraints, one per line,
each of the following form:
\code{seq1_idx seq1_start seq1_end seq2_idx seq2_start seq2_end}.
The COBALT \code{-c} flag}

\item{conserved_domain_threshold}{the threshold of the E-value
for selecting conserved domains.
The COBALT \code{-evalue} flag}

\item{filler_threshold}{the threshold of the E-value
for aligning filler segments.
The COBALT \code{-evalue2} flag}

\item{terminal_open_gap_penalty}{the gap open penalty
for initial/terminal gaps.
The COBALT \code{-g0} flag}

\item{terminal_extend_gap_penalty}{The COBALT \code{-e0} flag}

\item{open_gap_in_middle_penalty}{The COBALT \code{-g1} flag}

\item{middle_extend_gap_penalty}{The COBALT \code{-e1} flag}

\item{conservation_score_threshold}{the minimum average score needed
for a multiple alignment column to be considered as conserved.
The COBALT \code{-ccc} flag}

\item{dfb_added_weight}{the amount of extra weight (0..1)
to give the actual sequence letter at that position,
The COBALT \code{-dfb} flag}

\item{ffb_added_weight}{the amount of extra weight
(0..1) to give the actual sequence letter at that position
when assigning filler residue frequencies.
The COBALT \code{-ffb} flag}

\item{no_rps}{do not perform initial RPS blast search
The COBALT \code{-norps} flag}

\item{search_around_conserved_columns}{look for conserved columns
and iterate if any are found.
The COBALT \code{-iter} flag}

\item{score_matrix_name}{name of the COBALT score matrix.
Use \link{get_matrix_names} to get a list of all possible matrix names}

\item{pseudocount_constant}{the pseudocount constant
The COBALT \code{-pseudo} flag}

\item{use_fastme}{use FastME tree generation algorithm (instead
of neighbor joining).
The COBALT \code{-fastme} flag}

\item{do_dry_run}{Do a dry run, that is, only test all preconditions
The COBALT \code{-dryrun} flag}

\item{verbose}{verbose output
the COBALT \code{-v} option}
}
\description{
Create the COBALT options
}
