% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_default_cobalt_folder.R
\name{get_default_cobalt_folder}
\alias{get_default_cobalt_folder}
\title{Get the default COBALT folder}
\usage{
get_default_cobalt_folder()
}
\value{
the default COBALT folder
}
\description{
Get the default COBALT folder
}
