% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_cobalt.R
\name{install_cobalt}
\alias{install_cobalt}
\title{Install COBALT}
\usage{
install_cobalt(cobalt_folder = get_default_cobalt_folder())
}
\arguments{
\item{cobalt_folder}{the folder where COBALT is installed in.
Use \link{get_default_cobalt_folder} to get the default COBALT folder}
}
\description{
Install COBALT
}
