% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_cobalt_installed.R
\name{is_cobalt_installed}
\alias{is_cobalt_installed}
\title{See if COBALT is installed}
\usage{
is_cobalt_installed(
  cobalt_folder = get_default_cobalt_folder(),
  verbose = FALSE
)
}
\arguments{
\item{cobalt_folder}{the folder where COBALT is installed in.
Use \link{get_default_cobalt_folder} to get the default COBALT folder}

\item{verbose}{if TRUE, additional information is displayed, that
is potentially useful in debugging}
}
\description{
See if COBALT is installed
}
