% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_cobalt.R
\name{run_cobalt}
\alias{run_cobalt}
\title{Run COBALT}
\usage{
run_cobalt(
  fasta_filename,
  matrix_name = "BLOSUM62",
  cobalt_folder = get_default_cobalt_folder(),
  verbose = FALSE
)
}
\arguments{
\item{fasta_filename}{name of a FASTA file}

\item{matrix_name}{name of the COBALT matrix.
Use \link{get_matrix_names} to get a list of all possible matrix names}

\item{cobalt_folder}{the folder where COBALT is installed in.
Use \link{get_default_cobalt_folder} to get the default COBALT folder}

\item{verbose}{if TRUE, additional information is displayed, that
is potentially useful in debugging}
}
\value{
text of a FASTA file
}
\description{
Will stop if COBALT has an error.
}
