% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_cobalt_help.R
\name{run_cobalt_help}
\alias{run_cobalt_help}
\title{Show the COBALT help page}
\usage{
run_cobalt_help(cobalt_folder = get_default_cobalt_folder())
}
\arguments{
\item{cobalt_folder}{the folder where COBALT is installed in.
Use \link{get_default_cobalt_folder} to get the default COBALT folder}
}
\description{
Show the COBALT help page
}
