% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uninstall_cobalt.R
\name{uninstall_cobalt}
\alias{uninstall_cobalt}
\title{Uninstall COBALT}
\usage{
uninstall_cobalt(cobalt_folder = get_default_cobalt_folder())
}
\arguments{
\item{cobalt_folder}{the folder where COBALT is installed in.
Use \link{get_default_cobalt_folder} to get the default COBALT folder}
}
\description{
Uninstall COBALT
}
