% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_cobalt_installation.R
\name{check_cobalt_installation}
\alias{check_cobalt_installation}
\title{Check that COBALT is installed}
\usage{
check_cobalt_installation(cobalt_folder = get_default_cobalt_folder())
}
\arguments{
\item{cobalt_folder}{the folder where COBALT is installed in.
Use \link{get_default_cobalt_folder} to get the default COBALT folder}
}
\description{
Check that COBALT is installed, give a helpful error message if not
}
