% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cobaltr_self_test.R
\name{cobaltr_self_test}
\alias{cobaltr_self_test}
\title{Self-test the package}
\usage{
cobaltr_self_test(
  cobalt_options = create_cobalt_options(),
  cobalt_folder = get_default_cobalt_folder()
)
}
\arguments{
\item{cobalt_options}{a set of COBALT options,
as can be created by \link{create_cobalt_options}}

\item{cobalt_folder}{the folder where COBALT is installed in.
Use \link{get_default_cobalt_folder} to get the default COBALT folder}
}
\description{
Self-test the package
}
\examples{
if (is_cobalt_installed()) {
  cobaltr_self_test()
}
}
\author{
Richèl J.C. Bilderbeek
}
