% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_cobalt_installed.R
\name{is_cobalt_installed}
\alias{is_cobalt_installed}
\title{See if COBALT is installed}
\usage{
is_cobalt_installed(
  cobalt_folder = file.path(rappdirs::user_data_dir(), "cobalt"),
  verbose = FALSE
)
}
\description{
See if COBALT is installed
}
