% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_cobalt.R
\name{run_cobalt}
\alias{run_cobalt}
\title{Run COBALT}
\usage{
run_cobalt(
  fasta_filename,
  cobalt_folder = get_default_cobalt_folder(),
  verbose = FALSE
)
}
\arguments{
\item{fasta_filename}{name of a FASTA file}

\item{cobalt_folder}{the folder where COBALT is installed in.
Use \link{get_default_cobalt_folder} to get the default COBALT folder}

\item{verbose}{if TRUE, additional information is displayed, that
is potentially useful in debugging}
}
\description{
Run COBALT
}
