% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_cobalt.R
\name{run_cobalt}
\alias{run_cobalt}
\title{Run COBALT}
\usage{
run_cobalt(
  fasta_filename,
  cobalt_options = create_cobalt_options(),
  cobalt_folder = get_default_cobalt_folder(),
  verbose = FALSE
)
}
\arguments{
\item{fasta_filename}{name of a FASTA file}

\item{cobalt_options}{a set of COBALT options,
as can be created by \link{create_cobalt_options}}

\item{cobalt_folder}{the folder where COBALT is installed in.
Use \link{get_default_cobalt_folder} to get the default COBALT folder}

\item{verbose}{verbose output
the COBALT \code{-v} option}
}
\value{
text of a FASTA file
}
\description{
Will stop if COBALT has an error.
}
