% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_cobalt_help.R
\name{run_cobalt_help}
\alias{run_cobalt_help}
\title{Show the COBALT help page}
\usage{
run_cobalt_help(cobalt_folder = file.path(rappdirs::user_data_dir(), "cobalt"))
}
\description{
Show the COBALT help page
}
